Dungeons and Threats final iteration
Tomorrow the skills and stat gains will be adjusted. In addition the last of the skills will be added as well as the unique item for the mummy boss (I would do it this evening but I'm sure it would break something lol).

Progress has been slow with moving. Its about to speed back up.